Implemeted: -user can create a new order -add menu item to order -remove menu item from order -view order items -place order -view profile info -view restaurant info -view order history: ongoing and completed orders -login -register -logout

+++ order-mgmt/src/main/kotlin/com/order/ordermgmt/models/GroupOrder.kt
@@ -9,7 +9,7 @@ class GroupOrder(var orderStatus:OrderStatus):BaseEntity(){
     @JsonIgnoreProperties("groupOrder")
     val clientOrderList= mutableListOf<ClientOrder>()
 
-    var totalPrice:Double=0.0
+    var totalPrice:Int=0
 
     fun addClientOrder(clientOrder: ClientOrder){
         clientOrderList.add(clientOrder)
